+++ RELEASE_NOTES.md
@@ -352,3 +352,4 @@
 - Added function `iot_data_binary_from_string` to create binary data from string data
 - Added function `iot_data_map_get_array` to find an array in a map
 - Added function `iot_data_string_from_binary` to create string data from binary data
+- Added function `iot_data_binary_take` to take buffer from binary data